+++ src/adapters/pi/mcp-bridge.ts
